Initialize timers before starting the LoRa thread
When the first LoRa join request failed, the device was left frozen with
a turned green LED. Even the USB shell was not responsive (no echo
characters seen, no commands executed).

The LoRa thread for handling JOINs is using a timer. But the timers were
initialized _after_ that thread was run.

Signed-off-by: Dimitar Dimitrov <dimitar@dinux.eu>